+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Decouple from the `org.realityforge.gwt.keycloak:gwt-keycloak` dependency.
+
 ### [v6.183](https://github.com/replicant4j/replicant/tree/v6.183) (2026-03-30) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.182...v6.183)
 
 Changes in this release: